Allow immediates as operands in ZKIR v3 instructions (#36)
They are represented as hex strings in the JSON output.  If a string begins with "0x" it is an immediate.
The `load_imm` instruction is kept in case we want to write an immediate to a specific variable, but it's not emitted in the tests.
The `copy` instruction only works for variable to variable copies, not immediates.  We could change that.

There is still some cleanup to do.  We can probalby get rid of `Triv`, threading the instruction stream in a lot of places,
and we probalby don't need the impact operand datatype (since symbols and numbers are disjoint).